+++ testApi/src/main/scala/africa/absa/testing/scapi/rest/response/action/AssertionResponseAction.scala
@@ -18,7 +18,7 @@ package africa.absa.testing.scapi.rest.response.action
 
 import africa.absa.testing.scapi.json.ResponseAction
 import africa.absa.testing.scapi.logging.Logger
-import africa.absa.testing.scapi.rest.response.`enum`.AssertResponseActionType._
+import africa.absa.testing.scapi.rest.response.action.types.AssertResponseActionType._
 import africa.absa.testing.scapi.rest.response.Response
 import africa.absa.testing.scapi.utils.validation.ContentValidator
 import africa.absa.testing.scapi.{AssertionException, UndefinedResponseActionTypeException}
